Final touches, docstrings, nfo-s, etc.
