+++ car/car.nfo
@@ -1,3 +1,5 @@
 # Behold! The car's sources shall be uploaded to this folder.
 
 Orange Pi Zero with ARMbian is the current platform.
+
+- tcp_car.py is the main car script, which coordinates connecting to and streaming to a server.
