+++ generic/const.py
@@ -9,8 +9,3 @@ TICK = .1
 
 # Standard RGB data type, 0-255 unsigned int
 DTYPE = np.uint8
-
-# Addresses
-BEAST = "192.168.1.2"
-NOTE = "192.168.1.5"
-PI = "192.168.1.121"
